fixed the results page, its showing personalised tips..
