+++ Templates/contentpage.html
@@ -1,3 +1,5 @@
+{% include 'partials/chat.html' %}
+
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -160,6 +162,7 @@
             </div>
         </div>
     </div>
+    {% include 'partials/chat.html' %}
 
     <div class="modal-overlay" id="overlay" onclick="closeModal()"></div>
