+++ Templates/signup.html
@@ -13,10 +13,10 @@
 
         body {
             background-color: #e6c3f0;
-            color: #FFDA63;
+            color: #11100e;
             font-family: 'Lato', sans-serif;
             margin: 0;
-            padding: 0;
+            padding: 0; 
             display: flex;
             justify-content: center;
             align-items: center;
@@ -192,7 +192,7 @@
 
         form+p a {
             text-decoration: none;
-            color: #FFEB3B;
+            color: #11100e;
         }
 
         .signup-options {
@@ -279,6 +279,16 @@
             color: #dcdad3;
             font-size: 14px;
         }
+         
+::-webkit-scrollbar {
+    display: none; /* Hides the scrollbar in WebKit browsers */
+}
+
+html, body {
+    scrollbar-width: none; /* Hides the scrollbar in Firefox */
+    -ms-overflow-style: none; /* Hides the scrollbar in Internet Explorer and Edge */
+}
+
     </style>
 </head>
 
@@ -297,7 +307,7 @@
             <input type="password" id="password" name="password" placeholder="Enter your password" required />
             <label for="confirm-password">Confirm Password</label>
             <input type="password" id="confirm-password" placeholder="Confirm your password" required />
-            <input type="submit" value="Create your TVNerd account">
+            <input type="submit" value="Create your EsylianWell account">
         </form>
