--- index.css
@@ -1,142 +1,0 @@
-body {
-    background-color: rgba(190, 164, 236, 1);
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    text-align: center;
-    overflow-x: hidden;
-    
-}
-.header {
-    width: 100%;
-    padding: 10px;
-    display: flex;
-    justify-content: flex-end;
-    background-color: #BEAEDB;
-    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-}
-.header a {
-    /* margin-left: 20px; */
-    text-decoration: none;
-    color: #4E2487;
-    font-weight: bold;
-    transition: color 0.3s;
-    margin-right:20px;
-}
-.header a:hover {
-    color:#4B0082;
-}
-.container {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    animation: fadeIn 1.5s ease-in-out;
-    gap: 10px;
-}
-@keyframes fadeIn {
-    from { opacity: 0; }
-    to { opacity: 1; }
-}
-.title {
-    font-size: 60px;
-    font-weight: bold;
-    color: #9932CC;
-    text-transform: uppercase;
-    margin-bottom: 10px;
-    animation: slideDown 1s ease-in-out;
-    margin-top: 200px;
-    z-index: 100; /* Increased z-index */
-    position: relative; /* Ensure z-index works */
-    font-family: 'Poppins', sans-serif;
-    font-weight:700;
-    margin-top:100px;
-}
-@keyframes slideDown {
-    from { transform: translateY(-50px); opacity: 0; }
-    to { transform: translateY(0); opacity: 1; }
-}
-.description {
-    font-size: 22px;
-    color: #7658B2;
-    max-width: 60%;
-    animation: fadeIn 2s ease-in-out;
-    font-family: 'Poppins', sans-serif;
-    font-weight:400;
-}
-.box {
-    display: flex;
-    justify-content: center;
-    gap: 20px;
-    margin-top: 20px;
-}
-.button-box {
-    width: 150px; /* Increased width */
-    height: 150px; /* Increased height */
-    display: flex;
-    flex-direction: column; /* Added to stack inputs vertically */
-    justify-content: center;
-    align-items: center;
-    background-color: #B68EEA;
-    border-radius: 10px;
-    box-shadow: 2px 2px 10px rgba(0,0,0,0.2);
-    font-size: 18px;
-    font-weight: bold;
-    cursor: pointer;
-    transition: transform 0.3s, background-color 0.3s, box-shadow 0.3s;
-    animation: fadeIn 1.5s ease-in-out;
-    color:#592FA3;
-}
-.button-box input {
-    margin: 5px 0; /* Added margin for spacing between inputs */
-    padding: 5px;
-    font-size: 14px;
-}
-.button-box:hover {
-    transform: scale(1.1);
-    background-color: #A849C5;
-    box-shadow: 5px 5px 15px rgba(0,0,0,0.3);
-}
-.footer {
-    margin-top: 170px;
-    font-size: 18px;
-    color: #444;
-    animation: fadeIn 3s ease-in-out;
-}
-
-/* Styles for video icons */
-.video-container {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    pointer-events: none;
-}
-.video-icon {
-    position: absolute;
-    width: 100px; /* Increased width */
-    height: 100px; /* Increased height */
-    border-radius: 50%;
-    object-fit: cover;
-}
-.video-icon.left {
-    animation: moveLeft 10s infinite;
-}
-.video-icon.right {
-    right: 0; /* Position the icon to the right */
-    animation: moveRight 10s infinite;
-}
-
-/* Additional styles for date box */
-.hidden {
-    display: none;
-}
-#date-box.expanded {
-    width: 300px; /* Adjust width when expanded */
-    height: 300px; /* Adjust height when expanded */
-    transition: width 0.3s, height 0.3s; /* Smooth transition */
-}
