--- index.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Women's Health</title>
-    <link rel="stylesheet" href="/static/css/index.css">
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;600&family=Poppins:wght@600;700&display=swap" rel="stylesheet">
-</head>
-<body>
-    <div class="header">
-        <a href="/templates/login.html">Login</a>
-        <a href="/templates/signup.html">Sign Up</a>
-    </div>
-    <div class="container">
-        <div class="title">Elysian Well</div>
-        <div class="description">Provides comprehensive resources and support for mental health, 
-            reproductive care, and overall well-being, including guidance, educational materials, 
-            professional assistance, and community programs to promote a healthier and more balanced life.
-
-        </div>
-
-        <div class="box">
-            <div class="button-box" onclick="location.href='/templates/symptoms.html'">Symptoms</div>
-            <div class="button-box" id="date-box" onclick="toggleDateInputs(event)">
-                <span id="date-box-text">Add Dates</span>
-                <div id="date-inputs" class="hidden">
-                    <label for="start-date">Start Date:</label>
-                    <input type="date" id="start-date" name="start-date"> <br>
-                    <label for="end-date">End Date:</label>
-                    <input type="date" id="end-date" name="end-date">
-                </div>
-            </div>
-            <div class="button-box" onclick="location.href='/templates/Period guide and Info on irregular period.html'">Educational Content</div>
-        </div>
-      
-        <div class="footer">&copy; 2025 Women's Health Awareness</div>
-    </div>
-
-    <script src="/static/javascript/index.js"></script>
-</body>
-</html>
