+++ static/css/index.css
@@ -1,29 +1,60 @@
 body {
-    background-color: rgba(190, 164, 236, 1);
-    font-family: Arial, sans-serif;
+    font-family: 'Inter', sans-serif;
+    background-color: #f5f5f5;
     margin: 0;
     padding: 0;
     text-align: center;
     overflow-x: hidden;
 }
+
+/* Header */
 .header {
-    width: 100%;
-    padding: 10px;
     display: flex;
     justify-content: flex-end;
+    padding: 20px;
     background-color: #fff;
-    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
+
 .header a {
-    margin-left: 20px;
     text-decoration: none;
-    color: #444;
+    color: #4E2487;
     font-weight: bold;
-    transition: color 0.3s;
+    font-size: 1.1rem;
+    padding: 8px 15px;
+    border-radius: 8px;
+    transition: all 0.3s ease-in-out;
+    position: relative;
+    overflow: hidden;
+    margin-right: 20px;
 }
+
+.header a:last-child {
+    margin-right: 50px;
+}
+
+.header a::after {
+    content: "";
+    position: absolute;
+    bottom: 0;
+    left: 50%;
+    width: 0;
+    height: 2px;
+    background-color: #4B0082;
+    transition: width 0.3s ease-in-out, left 0.3s ease-in-out;
+}
+
+.header a:hover::after {
+    width: 100%;
+    left: 0;
+}
+
 .header a:hover {
-    color: #f08080;
+    color: #36125E;
+    background: rgba(75, 0, 130, 0.1);
 }
+
+/* Main Container */
 .container {
     width: 100%;
     height: 100vh;
@@ -33,72 +64,97 @@ body {
     align-items: center;
     text-align: center;
     animation: fadeIn 1.5s ease-in-out;
+    gap: 10px;
 }
+
 @keyframes fadeIn {
     from { opacity: 0; }
     to { opacity: 1; }
 }
+
+/* Title */
 .title {
     font-size: 60px;
     font-weight: bold;
-    color: #fff;
+    color: #9932CC;
     text-transform: uppercase;
     margin-bottom: 10px;
     animation: slideDown 1s ease-in-out;
-    margin-top: 200px;
-    z-index: 100; /* Increased z-index */
-    position: relative; /* Ensure z-index works */
+    z-index: 100;
+    position: relative;
+    font-family: 'Poppins', sans-serif;
+    font-weight: 700;
+    margin-top: 100px;
 }
+
 @keyframes slideDown {
     from { transform: translateY(-50px); opacity: 0; }
     to { transform: translateY(0); opacity: 1; }
 }
+
+/* Description */
 .description {
     font-size: 22px;
-    color: #222;
+    color: #7658B2;
     max-width: 60%;
     animation: fadeIn 2s ease-in-out;
+    font-family: 'Poppins', sans-serif;
+    font-weight: 400;
 }
+
+/* Buttons */
 .box {
     display: flex;
     justify-content: center;
     gap: 20px;
     margin-top: 20px;
 }
+
 .button-box {
-    width: 150px; /* Increased width */
-    height: 150px; /* Increased height */
+    width: 180px;
+    height: 60px;
     display: flex;
-    flex-direction: column; /* Added to stack inputs vertically */
     justify-content: center;
     align-items: center;
-    background-color: #fff;
-    border-radius: 10px;
-    box-shadow: 2px 2px 10px rgba(0,0,0,0.2);
-    font-size: 18px;
+    background: linear-gradient(135deg, #6A11CB, #A569BD);
+    border-radius: 30px;
+    box-shadow: 3px 3px 10px rgba(0,0,0,0.2);
+    font-size: 20px;
     font-weight: bold;
     cursor: pointer;
-    transition: transform 0.3s, background-color 0.3s, box-shadow 0.3s;
-    animation: fadeIn 1.5s ease-in-out;
+    transition: all 0.3s ease-in-out;
+    color: white;
+    text-transform: uppercase;
+    border: none;
+    text-align: center;
+    padding: 10px 20px;
+    text-decoration: none;
 }
-.button-box input {
-    margin: 5px 0; /* Added margin for spacing between inputs */
-    padding: 5px;
-    font-size: 14px;
-}
+
 .button-box:hover {
     transform: scale(1.1);
-    background-color: rgba(240, 128, 128, 0.8);
+    background: linear-gradient(135deg, #5A0FC8, #9B59B6);
     box-shadow: 5px 5px 15px rgba(0,0,0,0.3);
 }
+
+.button-box input {
+    margin: 5px 0;
+    padding: 5px;
+    font-size: 14px;
+    border: none;
+    border-radius: 5px;
+    text-align: center;
+}
+
+/* Footer */
 .footer {
-    margin-top: 400px;
+    margin-top: 170px;
     font-size: 18px;
     color: #444;
     animation: fadeIn 3s ease-in-out;
 }
 
-/* Styles for video icons */
+/* Video Icons */
 .video-container {
     position: absolute;
     top: 0;
@@ -107,20 +163,25 @@ body {
     height: 100%;
     pointer-events: none;
 }
+
 .video-icon {
     position: absolute;
-    width: 100px; /* Increased width */
-    height: 100px; /* Increased height */
+    width: 80px;
+    height: 80px;
     border-radius: 50%;
     object-fit: cover;
+    transition: transform 0.3s ease-in-out;
 }
+
 .video-icon.left {
     animation: moveLeft 10s infinite;
 }
+
 .video-icon.right {
-    right: 0; /* Position the icon to the right */
+    right: 0;
     animation: moveRight 10s infinite;
 }
+
 @keyframes moveLeft {
     0% { transform: translate(0, 0); }
     25% { transform: translate(-50px, 150px); }
@@ -128,6 +189,7 @@ body {
     75% { transform: translate(-150px, 200px); }
     100% { transform: translate(0, 0); }
 }
+
 @keyframes moveRight {
     0% { transform: translate(0, 0); }
     25% { transform: translate(100px, 150px); }
@@ -136,12 +198,81 @@ body {
     100% { transform: translate(0, 0); }
 }
 
-/* Additional styles for date box */
+/* Date Box */
 .hidden {
     display: none;
 }
+
 #date-box.expanded {
-    width: 300px; /* Adjust width when expanded */
-    height: 300px; /* Adjust height when expanded */
-    transition: width 0.3s, height 0.3s; /* Smooth transition */
+    width: 300px;
+    height: 300px;
+    transition: width 0.3s, height 0.3s;
+}
+
+/* Dialog */
+.dialog {
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    width: 300px;
+    background-color: white;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+    border-radius: 8px;
+    z-index: 1000;
+    display: none;
+}
+
+.dialog-content {
+    padding: 20px;
+    text-align: center;
+}
+
+.close-button {
+    position: absolute;
+    top: 10px;
+    right: 10px;
+    font-size: 20px;
+    cursor: pointer;
+}
+
+.dialog.show {
+    display: block;
+}
+
+#calendar-popup {
+    background-color: white;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+    margin-top: 20px;
+}
+
+#calendar-popup label {
+    display: block;
+    margin-bottom: 10px;
+    font-weight: bold;
+}
+
+#calendar-popup input {
+    display: block;
+    margin-bottom: 20px;
+    padding: 10px;
+    width: 100%;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+#calendar-popup button {
+    background-color: #6A11CB;
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+#calendar-popup button:hover {
+    background-color: #5A0FC8;
 }
