--- take3/templates/index.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Women's Health</title>
-    <link rel="stylesheet" href="/static/css/index.css">
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;600&family=Poppins:wght@600;700&display=swap" rel="stylesheet">
-</head>
-<body>
-    <div class="header">
-        {% if user %}
-            <span>Hi, {{ user.username }}</span>
-            <a href="{{ url_for('logout') }}">Logout</a>
-        {% else %}
-            <a href="{{ url_for('login') }}">Login</a>
-            <a href="{{ url_for('signup') }}">Sign Up</a>
-        {% endif %}
-    </div>
-    <div class="container">
-        <div class="title">Esylian Well</div>
-        <div class="description">
-            Provides comprehensive resources and support for mental health, 
-            reproductive care, and overall well-being, including guidance, educational materials, 
-            professional assistance, and community programs to promote a healthier and more balanced life.
-        </div>
-
-        <div class="box">
-            <div class="button-box" onclick="location.href='{{ url_for('symptoms') }}'">Symptoms</div>
-            <div class="button-box" id="calendar-box" onclick="toggleCalendar(event)">
-                <span id="calendar-box-text">Predict Next Period</span>
-            </div>
-            <div class="button-box" onclick="location.href='{{ url_for('content') }}'">Education Content</div>
-        </div>
-
-        <!-- Calendar Popup -->
-        <div id="calendar-popup" class="hidden">
-            <label for="start-date">Start Date:</label>
-            <input type="date" id="start-date">
-            <label for="end-date">End Date:</label>
-            <input type="date" id="end-date">
-            <button onclick="calculateCycle()">Predict Next Period</button>
-        </div>
-
-        <!-- Prediction Dialog -->
-        <div id="prediction-dialog" class="dialog hidden">
-            <div class="dialog-content">
-                <span class="close-button" onclick="closeDialog()">&times;</span>
-                <div id="prediction-box"></div>
-            </div>
-        </div>
-
-        <div class="footer">&copy; 2025 Women's Health Awareness</div>
-    </div>
-
-    <script src="/static/js/index.js"></script>
-    
-    <!-- Chatbot Integration -->
-    {% include 'chat.html' %}
-</body>
-</html>
