--- take3/templates/login.html
@@ -1,133 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login</title>
-
-
-    <style>
-      @import url('https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700&display=swap');
-
-body {
-    background-color: #e6c3f0; /* Updated background */
-    color: #11100e;
-    font-family: 'Lato', sans-serif;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-#login-form-div {
-    background-color: #bb95bc; /* Updated form background */
-    width: 350px;
-    height: fit-content;
-    margin: 100px auto;
-    border: 1px solid #333333;
-    padding: 30px;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(255, 218, 99, 0.2);
-    text-align: center;
-}
-
-#login-form-div>h1 {
-    margin-bottom: 20px;
-    font-weight: 700;
-    font-size: 28px;
-    color: #ffc8ff;
-}
-
-input {
-    width: 100%;
-    height: 40px;
-    outline: none;
-    padding: 0 10px;
-    margin: 10px 0;
-    background-color: #ddb5dd;
-    border: 1px solid #4d4d4d;
-    border-radius: 4px;
-    color: black;
-    font-size: 14px;
-    transition: border-color 0.3s ease;
-}
-
-input:focus {
-    border-color: #FFDA63;
-}
-
-label {
-    font-size: 14px;
-    font-weight: 600;
-    color: #fff9fe;
-}
-
-form+p {
-    font-size: 14px;
-    line-height: 23px;
-    color: #FFDA63;
-}
-
-form+p a {
-    text-decoration: none;
-    color: #FFEB3B;
-    transition: color 0.3s ease;
-}
-
-form+p a:hover {
-    color: #FFDA63;
-}
-
-#create-account {
-    background-color: plum;
-    color: #000000;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    padding: 12px;
-    margin-top: 20px;
-    width: 100%;
-    font-size: 16px;
-    font-weight: bold;
-    transition: background-color 0.3s ease;
-}
-
-#create-account:hover {
-    background-color: rgb(163, 107, 163);
-}
-</style>
-
-</head>
-
-<body>
-
-    <div id="login-form-div">
-        <h1>Login</h1>
-        <form method="POST">
-            <label for="email">Email</label>
-            <input type="email" id="email" name="email" placeholder="Enter your email" required />
-            <!-- <label for="user">User Name</label>
-            <input type="text" id="user" name="username" placeholder="Enter username" required /> -->
-            <label for="password">Password</label>
-            <input type="password" id="password" name="password" placeholder="Enter your password" required />
-            <input type="submit" value="Login" />
-        </form>
-        <!-- <button id="create-account"  onclick="window.location.href='{{ url_for('signup') }}'" >Create a New Account</button>
-        <p>By logging in, you agree to Esylian Well's <a href="#">Conditions of Use</a> and <a href="#">Privacy Policy</a>.</p> -->
-        
-        <a href="{{ url_for('signup') }}">
-            <button id="create-account">Create a New Account</button>
-        </a>
-
-        <p>By logging in, you agree to Esylian Well's <a href="#">Conditions of Use</a> and <a href="#">Privacy Policy</a>.</p>
-        
-        <p>Don't have an account? <a href="{{url_for('signup')}}">Sign up</a></p>
-    </div>
-
-</body>
-
-</html>
